[Search] Start building UI for the book search screen
Summary:
Begin building out a simple UI to start with for the search screen. The primary
body of new code related to this new UI is the new `BookItem` composable. It is
comprised of a books cover image and it's associated metadata. This UI will
need to be refined in later commits.

- Rename the `Type` enum to `BookType`.
- Create a new `Cover` enum that will associate cover names to drawable
resources.
- Change the `GoodReadsRating.rating` from a Long to a Double.
- Move SearchScreen into it's own package called `search`.
- Create an empty`SearchViewModel`.