[Gson] Add deserialization capability with Gson
Summary:
The goal of this commit is to add the capability to deserialize the book data
found in the app's `books_information.json` raw resource file. To that end the
following additions and changes have been made.

New functionality:
- A new file called `ContextUtils` was created and one new Context extension
function was added, `Context.readRawResourceToString()`. This function takes a
resource id, in our case it will always be the books_information.json file,
opens it into an `InputStream`. Then attempts to read said streams output
`ByteArray` as a String and return it. This adds the capability to grab our
json data and retrieve it as a String.

- A singleton object `GsonHelper` was created to encapsulate logic related to
using Gson to serialize or deserialize data. One new function was added called
`deserializeBooksJson`, which adds capability to deserialize the books Json
specifically.

- A new wrapper class `Result` was also added. This is simply a wrapper that
allows the aforementioned data retrieval components to return either the data
on success or an exception if one occurs.

- A `getSearchScreenState()` method was added to the `SearchViewModel` that
utilizes the above code to read and deserialize the books data and return it as
a `SearchScreenState`. Note that in future commits this method should be
modified to use a coroutine to perform these potentially long running
operations.

Minor Changes:
- Various changes have been made to the `Book` class to allow Gson to
deserialize to it. Mostly just name changes for properties and a new `Chapter`
data class.